refactor: split the two worst-MI files (CodeHealth 86.6 -> 89.8)
Acting on the Code Health Dashboard's top Improve-&-ROI finding.

- weatherFilteringUtils.ts (MI 14.9, 536 LOC) -> split into weatherFilteringTypes
  (types/constants) + geoUtils (distance/bounds/sorting/stats) + the weather-filter
  logic. Now MI 20.8 (green), 266 LOC. Re-exports keep all callers unchanged.
- MapContainer.tsx: extract the self-contained showEndOfResultsNotification DOM
  toast into utils/mapNotifications. MI 15.1 -> 16.7 (still yellow — the rest is
  Leaflet effects, a separate hook-extraction left as a visually-validated follow-up).

Behavior-preserving: lint + type-check clean, full vitest suite green (incl. the
81 weather-filtering + 12 MapContainer tests). Refreshed code-health trend.
