+++ code-health/coupling-history.tsv
@@ -1,2 +1,2 @@
 date	modules	max_fanout	high_coupling_gt10	mean_instability_pct
-2026-06-30	60	13	1	30
+2026-06-30	63	13	1	31
