+++ code-health/duplication-history.tsv
@@ -1,2 +1,2 @@
 date	pct	clones	dup_lines
-2026-06-30	1.15	5	121
+2026-06-30	1.16	5	122
