+++ code-health/hotspot-history.tsv
@@ -1,2 +1,2 @@
 date	files	hotspots	top_file	top_score
-2026-06-30	56	14	apps/web/src/components/MapContainer.tsx	988
+2026-06-30	59	14	apps/web/src/components/MapContainer.tsx	962
