+++ code-health/hotspot-table.md
@@ -1,6 +1,6 @@
 | Score | Revisions | Cyclomatic | File |
 |--:|--:|--:|---|
-| 988 | 13 | 76 | [`apps/web/src/components/MapContainer.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/MapContainer.tsx) |
+| 962 | 13 | 74 | [`apps/web/src/components/MapContainer.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/MapContainer.tsx) |
 | 840 | 60 | 14 | [`apps/web/src/App.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/App.tsx) |
 | 300 | 12 | 25 | [`apps/web/src/hooks/usePOINavigation.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/hooks/usePOINavigation.ts) |
 | 272 | 8 | 34 | [`apps/web/src/components/ads/AdManager.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/ads/AdManager.tsx) |
