+++ code-health/maintainability-history.tsv
@@ -1,2 +1,2 @@
 date	files	mean_mi	median_mi	min_mi	green	yellow	red
-2026-06-30	56	37.6	36.4	14.9	53	3	0
+2026-06-30	59	37.9	36.4	16.7	57	2	0
